<fix>delete_tooltip_and_item_input_style
