+++ app/frontend/packs/application.js
@@ -21,6 +21,7 @@ import '../scripts/shared/itemIndex'
 import '../scripts/shared/itemEditForm'
 import '../scripts/shared/post'
 import '../scripts/shared/tooltip'
+import '../scripts/shared/input'
 import '../scripts/shared/mytask'
 import '../scripts/shared/githubLink'
 import 'scripts'
