+++ app/views/groups/index.html.erb
@@ -4,6 +4,7 @@
     <!-- <p class="title text-md main-board-title">看板：<%= @board.name %></p> -->
     <div class="my-3">
       <span class="subtitle my-2 ">主題列表 </span>
+      <input data-tooltip="123" placeholder="123456">
       <%= link_to '', new_board_group_path, class:"fas fa-plus-square text-lg rounded text-center" , style:"text-decoration: none; color: #59825A" %>
     </div>
     <div class="content">
@@ -49,7 +50,7 @@
                             <%= link_to '', item_path(item), method: "DELETE", data: { confirm: '您確定要刪除這個項目嗎?' }, class: "far fa-trash-alt", style:"text-decoration: none;" %>
                           </div>
                           
-                          <Input type="text" data-remote="true" class="item-name" data-item-id="<%= item.id %>" id="item-name-<%= item.id %>" data-item-name="<%= item.name %>" value="<%= item.name %>" style="background-color:#EFF4EF; outline:none;">
+                          <Input type="text" data-remote="true" class="item-name" data-item-id="<%= item.id %>" id="item-name-<%= item.id %>" data-item-name="<%= item.name %>" value="<%= item.name %>">
                         </td>
                         <td class="post-icon-and-counts" style="width: 50px;">
                           <%= link_to "", posts_item_path(item), remote: true, id:"post-icon-id-#{ item.id }" ,class:"postIcon far fa-comment-dots", style:"text-decoration: none;" %>
@@ -67,7 +68,7 @@
                           <% end %>
                         </td>
                         <td>
-                          <input type="text" class="item-description" data-item-id="<%= item.id %>" id="item-description-<%= item.id %>" data-item-description="<%= item.description %>" value="<%= item.description %>" style="background-color:#EFF4EF; outline:none;" >
+                          <input type="text" class="item-description" data-item-id="<%= item.id %>" id="item-description-<%= item.id %>" data-item-description="<%= item.description %>" value="<%= item.description %>">
                         </td>
                         <td class="item-point"><%= item.point %></td>                        
                         <% if item.due_date != nil && item.status != "已完成" %>
